requirements.txt
aiogram==2.25.2
openai
google-api-python-client
google-auth
apscheduler
